set up basics
