+++ src/app/models/hris/employee-qualifications.interface.ts
@@ -1,9 +1,10 @@
 export interface EmployeeQualifications {
     id : number;
-    highestQualification: string | null,
+    employeeId?: number,
+    highestQualification: number,
     school: string | null,
     degree: string | null,
     fieldOfStudy: string | null,
     yearObtained: string | null,
-    nqfLevel: string | null,
+    nqfLevel: number,
 }
